fix: correct the calculation for water volume change in tanks and avoid having singular coef matrix (mmt)
The change of the water volume in tanks is now calculated using a new method _calculate_tanks_dV

Some networks had zero length pipes which causing the soil rows in the coef matrix to be zero at the diagonal, resulting
in a singular matrix system. To fix this, the rows of all pipes of zero length are not modified. Thus, they will only
have 1 at the diagonal and zero elsewhere.

A Singular matrix system was formed also when users forget to input the Weather object even when there are nodes with
'soil' or 'air' wall boundary conditions. To fix this, the solver will create a Weather object and will use it without
changing its default values. A warning message is printed to notify the user about this action.